Moved the social dominance API into the SocialData module
It makes much more sense to split APIs by domain, than by service
providing them.
