+++ lib/sanbase_web/graphql/resolvers/social_data_resolver.ex
@@ -71,4 +71,24 @@ defmodule SanbaseWeb.Graphql.Resolvers.SocialDataResolver do
   def social_gainers_losers_status(_root, args, _resolution) do
     SocialData.social_gainers_losers_status(args)
   end
+
+  def social_dominance(
+        _root,
+        %{
+          slug: slug,
+          from: from,
+          to: to,
+          interval: interval,
+          social_volume_type: social_volume_type
+        },
+        _resolution
+      ) do
+    SocialData.social_dominance(
+      slug,
+      from,
+      to,
+      interval,
+      social_volume_type
+    )
+  end
 end
